scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
python-dateutil==2.8.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
deep-translator==1.11.4
pdfplumber==0.11.0
